requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[argon2]
bcrypt
# for jwt token generation and verification
python-jose[cryptography]